+++ CHANGES.md
@@ -3,13 +3,16 @@ Change Log
 
 ### MobX Development
 
+
 #### next release (8.0.0-alpha.59)
 * Update magda error message
 * Add a short report section if trying to view a `3d-tiles` item in a 2d map.
+* Fix bug in `Terria.interpretStartData`.
 * Add `ThreddsCatalogGroup` model.
 * Port `supportsColorScaleRange`, `colorScaleMinimum` and `colorScaleMaximimum` from `master` to `WebMapServiceCatalogItem` model.
 * Fix WMS legend for default styles.
 * Ported MapboxVectorTileCatalogItem ("mvt").
+* Request transparent legend from GeoServer.
 * [The next improvement]
 
 #### 8.0.0-alpha.58